Move platform-specific sections of search page to their respective components
This move adds the custom search hook for each platform to it's respective component, as well as the JSX handling the incoming data. No JSX has replaced the missing UI in search page at this stage.
